Calendar feature, fully integrated with main program and working now. Style refactored to be properly visible with distinct delete colour and some changing colours on hover to give feedback. Flask routes now using in session calendar_id variable which gets set on first calendar load after user_id session is checked. Need to pull user management changes and finish that integration next.

+++ cmpt370_project_user_management/MealPlanner.py
@@ -1,13 +1,15 @@
 from flask import Flask, render_template, request, redirect, url_for, session, jsonify
 import sqlite3
 import bcrypt
-
 from cmpt370_project_user_management.Model.calendar_service import CalendarService
 from db.setup_database import database_connection, create_tables, update_tables
+import os
 
 app = Flask(__name__)
 DB_NAME = "db/saucyapp.db"
 calendar_service = CalendarService()
+#Needed for session  - random key creation
+app.secret_key = os.urandom(24)
 
 # Routing to a home page - Randi.
 @app.route('/')
@@ -165,11 +167,11 @@ def add_recipe():
     # On GET, just show the form
     return render_template('recipe_add.html')
 
-#Helper method to create SQL database connection and cursor - Jordan
+#Helper method to create and return SQL database connection - Jordan
 def _get_connection():
     '''
     _get_connection function - helper to create a connection to SQL database
-    :return: connection: connection to SQL database
+    :return: connection: connection to SQL database listed in DB_NAME
     '''
     try:
         connection = sqlite3.connect(DB_NAME)
@@ -178,35 +180,6 @@ def _get_connection():
     except sqlite3.Error as e:
         print("Error connecting to database:", e)
 
-#Helper method to generate / get calendar_id - Jordan
-def _get_calendar(connection: sqlite3.Connection):
-    '''
-    get_calendar method - helper to get the correct calendar_id before loading the calendar view page
-                          returns existing id, creates new one for new user, or returns none if user_id not provided
-    :return: calendar_id for the in sesstion user_id
-             None if the user_id is None
-    Todo: Refactor into setting session "calendar_id" that can be checked in all functions rather than multiple queries
-    Todo: Should be done in user login system?, after user_id is added to session, immediately get and set calendar_id
-    '''
-
-    #user_id = session.get("user_id") - Todo: refactor to pull user information from session
-    user_id = 42  #Constant user_id for testing UI and database
-
-    if user_id is None:
-        return None
-    else:
-        calendar_id = calendar_service.create_new_calendar(connection, user_id)
-        return calendar_id
-@app.route('/calender_login', methods=['GET', 'POST'])
-def calender_login():
-    if request.method == 'POST':
-        user_id = 42
-        connection = _get_connection()
-        calendar_id = calendar_service.create_new_calendar(connection, user_id)
-        print("calendar_id: ",calendar_id)
-        connection.close()
-        
-
 #Calendar View route - Jordan
 @app.route('/calendar')
 def calendar_view():
@@ -214,22 +187,22 @@ def calendar_view():
 
     :return:
     """
-    # Open database connection with cursor
+    # Open database connection
     connection = _get_connection()
+    # user_id = session.get("user_id") - Todo: refactor to pull user information from session
+    user_id = 42  # Constant user_id for testing UI and database
 
-    #Todo: refactor to redirect to homepage if not in session, otherwise render calendar
-    '''if "user_id" not in session: 
-        return redirect(url_for("index"))
-    else:
-        calendar_id = get_calendar()
-        return render_template("calendar.html",calendar_id=calendar_id)
-    '''
     try:
-        #Constant calendar view for testing
-        calendar_id = _get_calendar(connection)
-        return render_template("calendar.html",calendar_id=calendar_id)
+        if user_id is None: #Todo: if "user_id" not in session:
+            print("user_id is None")
+            #redirect to homepage Todo: return redirect(url_for("homePage")
+        else:
+            calendar_id = calendar_service.create_new_calendar(connection, user_id)
+            print("Calendar Id:", calendar_id)
+            session["calendar_id"] = calendar_id
+            return render_template("calendar.html", calendar_id=calendar_id)
     except sqlite3.Error as e:
-        print("Error getting calendar:", e)
+        print("Error getting/creating calendar:", e)
     finally:
         #Close connection when done
         connection.close()
@@ -245,21 +218,23 @@ def api_events():
     connection = _get_connection()
 
     try:
-        # Get calendar_id from session user_id
-        calendar_id = _get_calendar(connection)
-        if calendar_id is None:
-            return jsonify({"error": "User not logged in"}), 401
+        # Get calendar_id from session
+        if "calendar_id" not in session:
+            print("Redirect for no calendar_id - api_events")
+            return redirect(url_for('home'))
+        calendar_id = session["calendar_id"]
         #Pull events from database
         events = calendar_service.get_calendar_events(connection,calendar_id)
         #Convert to FullCalendar JSON format
         full_calendar_events = []
         for event in events:
             full_calendar_events.append({
-                "id": event.getEventID(),
-                "title": event.getEventName(),
-                "start": event.getEventDate(),
+                "id": event["event_id"],
+                "title": event["event_name"],
+                "start": event["event_date"],
                 "extendedProps":{
-                    "timeSlot": event.getEventTime()
+                    "timeSlot": event["event_time"],
+                    "recipe_id": event["recipe_id"]
                 }
             })
         #return all calendar events
@@ -281,11 +256,14 @@ def api_add_event():
     # Open database connection with cursor
     connection = _get_connection()
     try:
+        if "calendar_id" not in session:
+            print("Redirect for no calendar_id - api_add_event")
+            return redirect(url_for('home'))
         data = request.json
         # Get calendar_id from session user_id
-        calendar_id = _get_calendar(connection)
-        if calendar_id is None:
-            return jsonify({"error": "no calendar id"}), 400
+        calendar_id = session["calendar_id"]
+        print("Calendar Id ADD:", calendar_id)
+        print("add request data:", data)
 
         #Create event using calendar service
         event_id = calendar_service.insert_calendar_event(
@@ -296,6 +274,7 @@ def api_add_event():
             event_date=data["event_date"],
             event_time=data["event_time"]
         )
+        print("Added Event Id:", event_id)
         return jsonify({"event_id": event_id})
     except sqlite3.Error as e:
         print("Error adding event:", e)
@@ -314,9 +293,11 @@ def api_delete_event(event_id):
     # Open database connection
     connection = _get_connection()
     try:
+        if "calendar_id" not in session:
+            print("Redirect for no calendar_id - api_delete_event")
+            return redirect(url_for('home'))
         deleted_event = calendar_service.delete_calendar_event(connection,event_id)
-        if deleted_event is None:
-            return jsonify({"event_deleted": False}), 404
+        print("Deleted Boolean:", deleted_event)
         return jsonify({"event_deleted": True})
     except sqlite3.Error as e:
         print("Error deleting event:", e)
@@ -332,19 +313,3 @@ if __name__ == '__main__':
         update_tables(conn)
         conn.close()
     app.run(debug=True)
-
-
-
-
-
-'''
-
-#Session reminder
-
-import os
-#Needed for session  - random key creation
-app.secret_key = os.urandom(24)
-#set session variables
-session["user_id"] = request.form.get("user_id") #or other method to get data
-        
-'''
